client/requirements.txt
pyserial==3.4
hexdump==3.3